Moved business logic from service to facade
