+++ src/main/java/cz/inspire/thesis/data/service/sport/sport/SportKategorieService.java
@@ -2,6 +2,7 @@ package cz.inspire.thesis.data.service.sport.sport;
 
 import cz.inspire.thesis.data.dto.sport.sport.SportKategorieDetails;
 import cz.inspire.thesis.data.dto.sport.sport.SportKategorieLocDetails;
+import cz.inspire.thesis.data.model.sport.sport.SportInstructorEntity;
 import cz.inspire.thesis.data.model.sport.sport.SportKategorieEntity;
 import cz.inspire.thesis.data.model.sport.sport.SportKategorieLocEntity;
 import cz.inspire.thesis.data.repository.sport.sport.SportKategorieRepository;
@@ -25,7 +26,7 @@ public class SportKategorieService {
     private SportKategorieRepository sportKategorieRepository;
 
     @Transactional
-    public String create(SportKategorieDetails details) throws CreateException {
+    public SportKategorieEntity create(SportKategorieDetails details) throws CreateException {
         try {
             SportKategorieEntity entity = new SportKategorieEntity();
             if (details.getId() == null) {
@@ -34,97 +35,21 @@ public class SportKategorieService {
             mapToEntity(entity, details);
 
             sportKategorieRepository.save(entity);
-            postCreate(details, entity);
-            return entity.getId();
+            return entity;
         } catch (Exception e) {
             throw new CreateException("Failed to create SportKategorie", e);
         }
     }
 
     @Transactional
-    public void postCreate(SportKategorieDetails details, SportKategorieEntity entity) throws CreateException {
+    public void save(SportKategorieEntity entity) throws ApplicationException {
         try {
-            if (details.getLocaleData() != null) {
-                for (SportKategorieLocDetails locDetails : details.getLocaleData().values()) {
-                    SportKategorieLocEntity locEntity = new SportKategorieLocEntity();
-                    locEntity.setId(generateGUID(locEntity));
-                    locEntity.setJazyk(locDetails.getJazyk());
-                    locEntity.setNazev(locDetails.getNazev());
-                    locEntity.setPopis(locDetails.getPopis());
-                }
-            }
-
-            if (details.getNadrazenaKategorieId() != null) {
-                SportKategorieEntity nadrazenaKategorie = sportKategorieRepository.findOptionalBy(details.getNadrazenaKategorieId())
-                        .orElseThrow(() -> new CreateException("Parent category not found"));
-                entity.setNadrazenaKategorie(nadrazenaKategorie);
-            }
-
             sportKategorieRepository.save(entity);
         } catch (Exception e) {
-            throw new CreateException("Failed to execute post-create logic for SportKategorie", e);
+            throw new ApplicationException("Failed while trying to save SportKategorieEntity ", e);
         }
     }
 
-    @Transactional
-    public void setDetails(SportKategorieDetails details) throws ApplicationException {
-        try {
-            SportKategorieEntity entity = sportKategorieRepository.findOptionalBy(details.getId())
-                    .orElseThrow(() -> new ApplicationException("SportKategorie entity not found"));
-
-            // Clear old locale data
-            entity.getLocaleData().clear();
-
-            if (details.getLocaleData() != null) {
-                for (SportKategorieLocDetails locDetails : details.getLocaleData().values()) {
-                    SportKategorieLocEntity locEntity = new SportKategorieLocEntity();
-                    locEntity.setId(generateGUID(locEntity));
-                    locEntity.setJazyk(locDetails.getJazyk());
-                    locEntity.setNazev(locDetails.getNazev());
-                    locEntity.setPopis(locDetails.getPopis());
-                }
-            }
-
-            // There is missing functionality in Bean for setting podrazenaKategorie for nadrazenKategorieEntity when creating new Entity with
-            // nadrazenaKategorie ... or is possible dead functionality and tree is not reversible, and entities only have parent not child/s
-            // nadrazenaKategorie.setPodrazenaKategorie(nadrazenaKategorie.getPodrazenzekategorie().add(entity);
-
-            if (details.getNadrazenaKategorieId() != null) {
-                SportKategorieEntity nadrazenaKategorie = sportKategorieRepository.findOptionalBy(details.getNadrazenaKategorieId())
-                        .orElseThrow(() -> new ApplicationException("Parent category not found"));
-                entity.setNadrazenaKategorie(nadrazenaKategorie);
-            } else {
-                entity.setNadrazenaKategorie(null);
-            }
-
-            sportKategorieRepository.save(entity);
-        } catch (Exception e) {
-            throw new ApplicationException("Failed to update SportKategorie", e);
-        }
-    }
-
-    public SportKategorieDetails getDetails(SportKategorieEntity entity) {
-        SportKategorieDetails details = new SportKategorieDetails();
-        details.setId(entity.getId());
-        details.setMultisportFacilityId(entity.getMultiSportFacilityId());
-        details.setMultisportServiceUUID(entity.getMultiSportServiceUUID());
-
-        if (entity.getNadrazenaKategorie() != null) {
-            details.setNadrazenaKategorieId(entity.getNadrazenaKategorie().getId());
-        }
-
-        Map<String, SportKategorieLocDetails> localeData = entity.getLocaleData().stream()
-                .collect(Collectors.toMap(SportKategorieLocEntity::getJazyk, loc -> {
-                    SportKategorieLocDetails locDetails = new SportKategorieLocDetails();
-                    locDetails.setJazyk(loc.getJazyk());
-                    locDetails.setNazev(loc.getNazev());
-                    locDetails.setPopis(loc.getPopis());
-                    return locDetails;
-                }));
-        details.setLocaleData(localeData);
-
-        return details;
-    }
 
     public List<SportKategorieEntity> findAll() {
         return sportKategorieRepository.findAll();
